libs/flask_core/requirements.txt
# WaddleBot Flask Core Library Dependencies
# ==========================================

# Core Framework
quart>=0.19.0
hypercorn>=0.16.0  # ASGI server for Quart
quart-flask-patch>=0.3.0  # Flask compatibility for Quart

# Database
pydal>=20240906.1
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Authentication & Authorization
flask-security-too>=5.4.0
authlib>=1.3.0
pyjwt>=2.8.0
bcrypt>=4.1.2

# Redis
redis>=5.0.1
hiredis>=2.3.2  # Optional C parser for better performance

# HTTP Client (async)
httpx>=0.26.0  # For async HTTP requests

# Data Validation
pydantic>=2.5.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
